requirements.txt
fastapi>=0.121.0
uvicorn[standard]>=0.38.0
numpy>=2.0.0
prometheus-client>=0.23.0
python-multipart>=0.0.20
numba-cuda[cu12]>=0.1.0

